citra_qt: Add a progress dialog
When the video dumping is to be stopped, the dumper has to wait for all the frames in the queue to be encoded. This can take a quite long time when the number of remaining frames is large, so I added a progress dialog about it.

Note, it does not really show "progress" currently (but rather just a text message)